refactor: adjust import paths in knowledge scripts to align with updated project structure
